Butter Charge attack Update - Ongoing
The charge attack for the butter enemy has been altered as such, if the player is aligned with the enemy on the x or y axis then the butter enemy will charge towards the furthest possible cell on the grid. The charge in practice is a bit jank and requires some fine tuning.